requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
black==25.1.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
docstring_parser==0.16
email_validator==2.2.0
fastapi==0.115.10
fastapi-cli==0.0.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
passlib3==0.1.0
pathspec==0.12.1
piccolo==1.23.0
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlmodel==0.0.23
starlette==0.46.0
targ==0.4.0
typer==0.15.2
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
